% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ch.est}
\alias{ch.est}
\title{Estimate \strong{s2},\strong{lambda2},\strong{lambda4},\strong{Delta}}
\usage{
ch.est(nu, gamma, size, B = 100)
}
\arguments{
\item{nu}{bandwidth of Gaussian kernal applied to White-noise, Whitenoise error if \code{nu} = 0}

\item{gamma}{bandwidth of nonparameter smoothing}

\item{size}{sample size}

\item{B}{Montelarlo iteration times}
}
\value{
a list of s2,lambda2,lambda4,Delta
}
\description{
Estimate \strong{s2},\strong{lambda2},\strong{lambda4},\strong{Delta}
}
\examples{
ch.est(nu=2,gamma=4,size=1000,B=100)
}
\references{
Multiple Testing of Local Extrema for Detection of Change Points \url{https://arxiv.org/abs/1504.06384}
}
\seealso{
\code{\link{which.cp}}
}
