% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Fdr}
\alias{Fdr}
\title{Evaluate performance of estimated change points}
\usage{
Fdr(uh, b, th)
}
\arguments{
\item{uh}{a vector of estimated change points locations}

\item{b}{a scalar of location tolerance, specified by user}

\item{th}{a vector of true change points locations}
}
\value{
a list of vector of \code{FDR} and \code{Power}
\item{FDR}{a scalar of fdr (false discovery rate)}
\item{Power}{a scalar of power (true positive rate)}
}
\description{
Evaluate performance of estimated change points
}
\examples{
Fdr(uh=c(7,15,32,47),b=4,th=c(10,20,30,40,50))
}
\seealso{
\code{\link{which.cp}}
}
