% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GenDY}
\alias{GenDY}
\title{Generate first-order differential of a smoothed sequence Y}
\usage{
GenDY(mu, z, gamma)
}
\arguments{
\item{mu}{a vector of piecewise constant}

\item{z}{a vector of stationary Gaussian random error}

\item{gamma}{bandwidth of nonparameter smoothing}
}
\value{
a vector of the differential of Y
}
\description{
Generate first-order differential of a smoothed sequence Y
}
\examples{
mu = GenMu(x=1:10,pos=seq(10,100,10),size=150)
z = GenZ(nu=2,size=150)
GenDY(mu=mu,z=z,gamma=4)
}
\seealso{
\code{\link{GenMu}}/\code{\link{GenZ}}
}
