% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GenMu}
\alias{GenMu}
\title{Generate a piecewise constant sequence starting from 0}
\usage{
GenMu(x, pos, size)
}
\arguments{
\item{x}{a vector containing all values of change points}

\item{pos}{positions of change points, corresponding to \code{x}}

\item{size}{sample size}
}
\value{
a piecewise constant sequence
}
\description{
Generate a piecewise constant sequence starting from 0
}
\examples{
GenMu(x=1:10,pos=seq(10,100,10),size=150)
}
\seealso{
\code{\link{GenDY}}
}
