% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GenZ}
\alias{GenZ}
\title{Generate Gaussian autocorrelated random error sequence based on White-noise and Gaussian kernal}
\usage{
GenZ(nu, size)
}
\arguments{
\item{nu}{bandwidth of Gaussian kernal applied to White-noise, Whitenoise error if \code{nu} = 0}

\item{size}{sample size}
}
\value{
a vector of random error
}
\description{
Generate Gaussian autocorrelated random error sequence based on White-noise and Gaussian kernal
}
\examples{
GenZ(nu=2,size=1000)
}
\seealso{
\code{\link{GenDY}}
}
